% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_true_false.R
\name{co_occurence_matrix}
\alias{co_occurence_matrix}
\title{Co-occurence matrix}
\usage{
co_occurence_matrix(TRUE_FALSE_DATAFRAME)
}
\arguments{
\item{TRUE_FALSE_DATAFRAME}{}
}
\value{

}
\description{
A useful function for calculated the co-occurences of dummy variables
}
