% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_true_false.R
\name{collapse_multi_column_dummy}
\alias{collapse_multi_column_dummy}
\title{Collapse dummy variables}
\usage{
collapse_multi_column_dummy(data_frame, columns_to_collapse)
}
\arguments{
\item{columns_to_collapse}{}
}
\value{

}
\description{
Reverse of List_to_True_False
}
