% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_true_false.R
\name{create_named_nested_list}
\alias{create_named_nested_list}
\title{Create a named nested list}
\usage{
create_named_nested_list(longer_list, shorter_list)
}
\arguments{
\item{shorter_list}{}
}
\value{

}
\description{
Will come back to this another time. I have completely forgotten what this function does, but I know it is necessary for the "list_to_true_false" function.
}
