% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{firstup}
\alias{firstup}
\title{A function for raising first letter to uppercase}
\usage{
firstup(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
A function for raising first letter to uppercase
}
