% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_true_false.R
\name{List_to_True_False}
\alias{List_to_True_False}
\title{List_to_True_False}
\usage{
List_to_True_False(x, seperator)
}
\arguments{
\item{x}{a col}

\item{seperator}{}
}
\value{

}
\description{
This is a function to convert categorical data from a single-column format a more usefule dummy-variable format
For example, in the single column format data would appear as follows:
    Months
1   jan mar ...
2   feb mar ...
3   mar ...
}
\details{
Using this function the output would be:
    Jan           Feb         Mar         ...
1   TRUE          FALSE       TRUE        ...
2   FALSE         TRUE        TRUE        ...
3   FALSE         FALSE       TRUE        ...
}
