% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{lower_quartile_function}
\alias{lower_quartile_function}
\title{Lower Quartile}
\usage{
lower_quartile_function(column)
}
\arguments{
\item{column}{}
}
\value{

}
\description{
Findthe lower quartile of a column
}
