% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{outlier_detection}
\alias{outlier_detection}
\title{Outlier Detection}
\usage{
outlier_detection(column)
}
\arguments{
\item{column}{Column to detect outliers}
}
\value{
A vector of booleans. Indicating which values are outliers and which values are nto
}
\description{
Outlier Detection
}
