% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_presentation_function.R
\name{presenting_info_for_table}
\alias{presenting_info_for_table}
\title{Summarise for table}
\usage{
presenting_info_for_table(column, seperator, percentage_cut_off)
}
\arguments{
\item{column}{The column you wish to summarise}

\item{seperator}{The string used to seperate the values in the string response}

\item{percentage_cut_off}{The cutoff you want to use. For example, the}
}
\value{
returns a string with the numbers summarised
}
\description{
Suitable for "column" in the format c("resp_a;resp_b", "resp_c;resp_a")
}
