% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_multiple_values.R
\name{replace_multiple_values}
\alias{replace_multiple_values}
\title{List must be written in form:
temp <- c("a"=10,"c"=20)
Where "a" and "c" are the original values}
\usage{
replace_multiple_values(original_list, conversion_list)
}
\arguments{
\item{original_list}{The list you would like to convert}

\item{conversion_list}{The named list you will use conversion}
}
\value{
Returns the converted list
}
\description{
Replace multiple values
}
\details{
A function for switching values in one list, using a named list
}
\examples{

}
