% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_multiple_values.R
\name{replace_multiple_values_exact}
\alias{replace_multiple_values_exact}
\title{Replace multiple values (exact)}
\usage{
replace_multiple_values_exact(original_list, conversion_list)
}
\arguments{
\item{original_list}{The list you would like to convert}

\item{conversion_list}{The named list you will use conversion}
}
\value{
Returns the converted list
}
\description{
A function for switching values in one list, using a named list. Useful in the case where you need exact conversion.
For example, if your list includes any values which are likely to overlap (e.g. man and woman.)
}
\examples{

}
