% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{summarise_indicator_sheet}
\alias{summarise_indicator_sheet}
\title{Summarise Indicator_sheet}
\usage{
summarise_indicator_sheet(indicator_sheet)
}
\arguments{
\item{indicator_sheet}{}
}
\value{

}
\description{
A function for summarising every single column of the indicator sheet
}
