% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality_checks.R
\name{upper_quartile_function}
\alias{upper_quartile_function}
\title{Upper quartile}
\usage{
upper_quartile_function(column)
}
\arguments{
\item{column}{}
}
\value{

}
\description{
A function for retrieving the upper quartile of a column
}
